moved functions from setup to dashboard
